feat: shrink dashboard frontend route shell, config pages, and interaction containers
